% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithm leveraging function}
\usage{
algo_leverage(y, X, size = NA, draws = NA, method = "weighted")
}
\arguments{
\item{y}{The constant vector}

\item{X}{The regressor}

\item{size}{Sample size of leveraging sampling.}

\item{draws}{Number of draws of subsample}

\item{method}{The method to implement, can be weighted leveraging and uniform
leveraging. Default is "weighted". Details of the difference of the}
}
\value{
The leverage solution to the linear regressioin problem.
}
\description{
A function that applies the method mentioned by (Ma, 2014) to do a leveraging
through two sampling methods, the uniform sampling and weighted sampling
}
\author{
Xiaohan Wang
}
