% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Net}
\usage{
elnet_coord(y, X, beta0 = NA, alpha = 0, niter = 10000, tol = 1e-05)
}
\arguments{
\item{y}{The constant vector}

\item{X}{The regressor}

\item{beta0}{The initial guess of the solution. If no input given, then it is
set to a zero vector.}

\item{alpha}{The elastic net parameter, alpha should be in [0,1]. More details
to be found in Trevors paper.}

\item{niter}{Maximum Number of Iteration}

\item{tol}{Tolerance of the difference between two result from iteration}
}
\value{
The solution to the elastic net system according to given parameter.
}
\description{
A function that implements the algorithm proposed by (Zou, Hui; Hastie,
Trevor, 2015, JRRSB).
}
\author{
Xiaohan Wang
}
