% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xw547_hw2.R
\docType{package}
\name{hw2}
\alias{hw2}
\title{xw547hw2: A package that integrates hw1 from BTRY6520}
\description{
This package gives a comprehensive integration of the existing functions of
the first homework.
}
\section{Functions}{

solve.ols, algo_leverage, elnet.coord
}

