% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xw547.R
\docType{package}
\name{overview}
\alias{overview}
\title{xw547: A package that integrates hw1 from BTRY6520}
\description{
This package gives a comprehensive integration of the existing functions of
the first homework.
}
\section{Functions}{

solve_ols, algo_leverage, elnet_coord
}

