% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Solve linear system}
\usage{
solve_ols(
  A,
  b,
  x0 = NA,
  maxiter = 10000,
  tol = 1e-05,
  numCores = 1,
  parallel = FALSE,
  method = "Gauss",
  full_result = FALSE
)
}
\arguments{
\item{A}{Linear System to be solved}

\item{b}{Constant Vector}

\item{x0}{Initial solution guess}

\item{maxiter}{Maximum Number of Iteration}

\item{tol}{Tolerance of the difference between two result from iteration}

\item{numCores}{Only applicable when parallel == TRUE, this}

\item{parallel}{Whether to use the paralleled version, Only available when
using "Jacobi". Default is FALSE.}

\item{method}{Method to use, default is "Gauss-Seidel", and can be "Jacobi"}

\item{full_result}{Whether to return a result of each iteration}
}
\value{
The function would return the solution to the linear system or the
full record of each iteration. This depends on the option of full_result.
}
\description{
solve.ols is a function that solves linear system using either Gauss-Seidel
or Jacobi. It assumes that the spectral norm of A is less than 1 to solve.
}
\author{
Xiaohan Wang
}
